Move hook tests from useGallery
