+++ packages/peregrine/lib/talons/Gallery/__tests__/useGallery.spec.js
@@ -4,13 +4,15 @@ import { MockedProvider } from '@apollo/client/testing';
 import { renderHook } from '@testing-library/react-hooks';
 
 import typePolicies from '../../../Apollo/policies';
-import operations from '../gallery.gql.ee';
-import {
-    mockGetWishlistConfig,
-    mockGetWishlistItemsPage1,
-    mockGetWishlistItemsPage2
-} from '../__fixtures__/apolloMocks';
+import { mockGetWishlistConfig } from '../__fixtures__/apolloMocks';
 import { useGallery } from '../useGallery';
+
+jest.mock(
+    '../../../hooks/useCustomerWishlistSkus/useCustomerWishlistSkus',
+    () => ({
+        useCustomerWishlistSkus: jest.fn()
+    })
+);
 
 jest.mock('../../../context/user', () => ({
     useUserContext: jest.fn().mockReturnValue([{ isSignedIn: true }])
@@ -22,12 +24,7 @@ const cache = new InMemoryCache({
 
 const renderHookWithProviders = ({
     renderHookOptions = {},
-    mocks = [
-        mockGetWishlistConfig,
-        mockGetWishlistItemsPage1,
-        mockGetWishlistItemsPage1,
-        mockGetWishlistItemsPage2
-    ]
+    mocks = [mockGetWishlistConfig]
 } = {}) => {
     const wrapper = ({ children }) => (
         <MockedProvider mocks={mocks} addTypename={true} cache={cache}>
@@ -56,25 +53,6 @@ test('returns store config', async () => {
             "id": 1,
             "magento_wishlist_general_is_enabled": "1",
           },
-        }
-    `);
-});
-
-test('pre-caches wishlist items', async () => {
-    const { waitForNextUpdate } = renderHookWithProviders();
-
-    await waitForNextUpdate();
-
-    const preCacheData = cache.readQuery({
-        query: operations.getProductsInWishlistsQuery
-    });
-
-    expect(preCacheData).toMatchInlineSnapshot(`
-        Object {
-          "customerWishlistProducts": Array [
-            "Dress",
-            "Shirt",
-          ],
         }
     `);
 });
